pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "manus-3subagent-repro"
version = "0.1.0"
description = "Reproduction scaffold for a Manus-style 3-subagent architecture"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langgraph>=0.2.0",
  "openai>=1.40.0",
  "pydantic>=2.8.0",
  "typer>=0.12.0",
  "pyyaml>=6.0.1",
  "orjson>=3.10.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
  "tenacity>=8.4.0",
  "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
]

[project.scripts]
manus3-run = "manus_three_agent.eval.runner:run_cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
